refactor: move tool_registry/plugin.rs to astra-turn-core (403 LOC)
Move the dynamic tool registration module to astra-turn-core where
TOOL_CATALOG already lives.

Changes:
- Copy plugin.rs to astra-turn-core::tool_registry_plugin (403 LOC)
- Register module in lib.rs
- Replace original with re-export stub
